+++ lib/src/components/shimmer.dart
@@ -37,17 +37,17 @@ class Shimmer extends StatefulWidget {
     this.loop = 0,
     this.enabled = true,
   }) : gradient = LinearGradient(
-         begin: Alignment.topLeft,
-         end: Alignment.centerRight,
-         colors: <Color>[
-           baseColor,
-           baseColor,
-           highlightColor,
-           baseColor,
-           baseColor,
-         ],
-         stops: const <double>[0.0, 0.35, 0.5, 0.65, 1.0],
-       );
+          begin: Alignment.topLeft,
+          end: Alignment.centerRight,
+          colors: <Color>[
+            baseColor,
+            baseColor,
+            highlightColor,
+            baseColor,
+            baseColor,
+          ],
+          stops: const <double>[0.0, 0.35, 0.5, 0.65, 1.0],
+        );
 
   @override
   State createState() => _ShimmerState();