+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* No major changes
+
+## 0.0.1
+
+* Added Creative Button V1
+* Added Async Button v1
+* Added Back Button v1